requirements.txt
selenium
requests
pillow
streamlit
tqdm